requirements.txt
# === 必须的核心运行依赖 ===
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3

# === LangChain + LangGraph 依赖（与你本地完全一致） ===
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1

# === LLM SDK ===
openai==2.7.2

# === 工具依赖 ===
typing_extensions==4.15.0
numpy==2.2.6